Fixed a compilation problem. I'm amazed no one else had this problem.

+++ community/kernel/src/main/java/org/neo4j/kernel/api/properties/ByteProperty.java
@@ -48,7 +48,7 @@ final class ByteProperty extends NumberPropertyWithin4Bytes
     {
         if ( other instanceof Byte )
         {
-            return other == (byte)value;
+            return ((Byte)other).byteValue() == value;
         }
 
         return valueCompare( value, other );
